Split SettingsStack
Create a `ReadOnlySettingsStack`, that has no writable support.
Annotabe both classes
